backend/requirements.txt
fastapi
uvicorn[standard]
supabase
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
python-dotenv
python-multipart
psycopg2-binary
pypdf
